- added highlighting of buildings within the influence radius - added drag-all mechanic by pressing both mouse buttons - added copy mechanic by double clicking - added delete button handling to delete selection - added a status message display - added more preset colors - added example layouts - improved icon rendering - improved UI menu and status bar

+++ trunk/AnnoDesigner/AnnoObject.cs
@@ -44,7 +44,7 @@ namespace AnnoDesigner
         /// Influence radius in grid units
         /// </summary>
         [DataMember]
-        public int Radius;
+        public double Radius;
 
         /// <summary>
         /// Empty constructor needed for deserialization
